+++ spec/grape/middleware/versioner/header_spec.rb
@@ -252,7 +252,7 @@ describe Grape::Middleware::Versioner::Header do
     end
   end
 
-  context 'when there are multiple versions specified with rescue_from :all' do
+  context 'when there are multiple versions with complex vendor specified with rescue_from :all' do
     subject {
       Class.new(Grape::API) do
         rescue_from :all
@@ -261,7 +261,11 @@ describe Grape::Middleware::Versioner::Header do
 
     let(:v1_app) {
       Class.new(Grape::API) do
-        version 'v1', using: :header, vendor: 'test'
+        version 'v1', using: :header, vendor: 'test.a-cool-resource'
+        content_type :v1_test, 'application/vnd.test.a-cool-resource-v1+json'
+        formatter :v1_test, ->(object, _) { object }
+        format :v1_test
+
         resources :users do
           get :hello do
             'one'
@@ -272,7 +276,11 @@ describe Grape::Middleware::Versioner::Header do
 
     let(:v2_app) {
       Class.new(Grape::API) do
-        version 'v2', using: :header, vendor: 'test'
+        version 'v2', using: :header, vendor: 'test.a-cool-resource'
+        content_type :v2_test, 'application/vnd.test.a-cool-resource-v2+json'
+        formatter :v2_test, ->(object, _) { object }
+        format :v2_test
+
         resources :users do
           get :hello do
             'two'
@@ -289,13 +297,13 @@ describe Grape::Middleware::Versioner::Header do
 
     context 'with header versioned endpoints and a rescue_all block defined' do
       it 'responds correctly to a v1 request' do
-        versioned_get '/users/hello', 'v1', using: :header, vendor: 'test'
+        versioned_get '/users/hello', 'v1', using: :header, vendor: 'test.a-cool-resource'
         expect(last_response.body).to eq('one')
         expect(last_response.body).not_to include('API vendor or version not found')
       end
 
       it 'responds correctly to a v2 request' do
-        versioned_get '/users/hello', 'v2', using: :header, vendor: 'test'
+        versioned_get '/users/hello', 'v2', using: :header, vendor: 'test.a-cool-resource'
         expect(last_response.body).to eq('two')
         expect(last_response.body).not_to include('API vendor or version not found')
       end